Adds VCS urls support.
Now, requirements, containing VCS urls could be
compiled. During compilation, pip-compile will checkout
sources, get all dependencies and pin spec to exact
repository revision.

+++ piptools/package_manager.py
@@ -21,11 +21,13 @@ from functools import partial
 import urlparse
 
 #from pip.backwardcompat import ConfigParser
-from pip.download import _download_url, _get_response_from_url
+from pip.download import _download_url, _get_response_from_url, unpack_vcs_link
 from pip.index import Link, PackageFinder, package_to_requirement
 #from pip.locations import default_config_file
 from pip.req import InstallRequirement
 from pip.util import splitext
+from pip.vcs import vcs
+
 
 from .logging import logger
 from .datastructures import Spec, first
@@ -55,7 +57,7 @@ class BasePackageManager(object):
         """
         raise NotImplementedError('Implement this in a subclass.')
 
-    def get_dependencies(self, name, version):
+    def get_dependencies(self, pinned_spec):
         """Return a list of Spec instances, representing the dependencies of
         the specific package version indicated by the args.  This method only
         returns the direct (next-level) dependencies of the package.
@@ -136,8 +138,8 @@ class FakePackageManager(BasePackageManager):
             raise NoPackageMatch('No package found for %s' % (spec,))
         return self.pick_highest(versions)
 
-    def get_dependencies(self, name, version):
-        pkg_key = '%s-%s' % (name, version)
+    def get_dependencies(self, pinned_spec):
+        pkg_key = '%s-%s' % (pinned_spec.name, pinned_spec.version)
         specs = []
         for specline in self._contents[pkg_key]:
             specs.append(Spec.from_line(specline))
@@ -246,7 +248,12 @@ class PackageManager(BasePackageManager):
                 link = self._link_cache[specline]
                 source = 'link cache'
             else:
-                requirement = InstallRequirement.from_line(specline)
+                if spec.url:
+                    # TODO : remove
+                    requirement = InstallRequirement.from_editable(specline)
+                else:
+                    requirement = InstallRequirement.from_line(specline)
+                    
                 finder = PackageFinder(
                     find_links=[],
                     index_urls=['https://pypi.python.org/simple/'],
@@ -260,22 +267,39 @@ class PackageManager(BasePackageManager):
 
             # Take this moment to smartly insert the pinned variant of this
             # spec into the link_cache, too
-            pinned_spec = Spec.from_pinned(spec.name, version)
+            pinned_spec = spec.pin(version)
             if pinned_spec not in self._link_cache:
                 self._link_cache[str(pinned_spec)] = link
             return version, source
 
         specline = str(spec)
-        if '==' not in specline or specline not in self._best_match_call_cache:
-            logger.debug('- Finding best package matching %s' % [specline])
-        with logger.indent():
-            version, source = _find_cached_match(spec)
-        if '==' not in specline or specline not in self._best_match_call_cache:
-            logger.debug('  Found best match: %s (from %s)' % (version, source))
+        if spec.url:
+            self._link_cache[specline] = Link(spec.vcs_url)
+            
+            path = self.get_or_download_package(spec)
+            version = self.get_vcs_revision(path)
+            pinned_spec = spec.pin(version)
+            self._link_cache[str(pinned_spec)] = Link(pinned_spec.vcs_url)
+        else:
+            if '==' not in specline or specline not in self._best_match_call_cache:
+                logger.debug('- Finding best package matching %s' % [specline])
+            with logger.indent():
+                version, source = _find_cached_match(spec)
+            if '==' not in specline or specline not in self._best_match_call_cache:
+                logger.debug('  Found best match: %s (from %s)' % (version, source))
+                
         self._best_match_call_cache[specline] = True
         return version
 
-    def get_dependencies(self, name, version):
+    def get_vcs_revision(self, path):
+        backend_cls = vcs.get_backend_from_location(path)
+        backend = backend_cls(path)
+        return backend.get_revision(path)
+    
+    def get_dependencies(self, pinned_spec):
+        name = pinned_spec.name
+        version = pinned_spec.version
+        
         key = '{0}-{1}'.format(name, version)
         if key not in self._dep_call_cache:
             logger.debug('- Getting dependencies for %s-%s' % (name, version))
@@ -284,8 +308,7 @@ class PackageManager(BasePackageManager):
             if deps is not None:
                 source = 'dependency cache'
             else:
-                spec = Spec.from_pinned(name, version)
-                path = self.get_or_download_package(str(spec))
+                path = self.get_or_download_package(pinned_spec)
                 deps = self.extract_dependencies(path)
                 self._dep_cache[(name, version)] = deps
                 source = 'package archive'
@@ -305,13 +328,13 @@ class PackageManager(BasePackageManager):
         fullpath = os.path.join(self.download_cache_root, cache_key)
         return fullpath
 
-    def get_or_download_package(self, specline):
+    def get_or_download_package(self, spec):
         """Returns the local path from the package cache, downloading as
         needed.
         """
-        logger.debug('- Getting package location for %s' % (specline,))
+        logger.debug('- Getting package location for %s' % (spec,))
         with logger.indent():
-            link = self._link_cache[str(specline)]
+            link = self._link_cache[str(spec)]
             fullpath = self.get_local_package_path(url_without_fragment(link))
 
             if os.path.exists(fullpath):
@@ -321,7 +344,13 @@ class PackageManager(BasePackageManager):
             logger.debug('  Archive cache miss, downloading {0}...'.format(
                 link.filename
             ))
-            return self.download_package(link)
+            if spec.url:
+                # get sources from vcs
+                unpack_vcs_link(link, fullpath, only_download=False)
+            else:
+                self.download_package(link, fullpath)
+
+            return fullpath
 
     # def get_pip_cache_root():
     #     """Returns pip's cache root, or None if no such cache root is
@@ -341,7 +370,7 @@ class PackageManager(BasePackageManager):
     #         download_cache = os.path.expanduser(download_cache)
     #     return download_cache
 
-    def download_package(self, link):
+    def download_package(self, link, destination):
         """Downloads the given package link contents to the local
         package cache. Overwrites anything that's in the cache already.
         """
@@ -357,10 +386,8 @@ class PackageManager(BasePackageManager):
         url = url_without_fragment(link)
         logger.debug('- Downloading package from %s' % (url,))
         with logger.indent():
-            fullpath = self.get_local_package_path(url)
             response = _get_response_from_url(url, link)
-            _download_url(response, link, fullpath)
-            return fullpath
+            _download_url(response, link, destination)
 
     def unpack_archive(self, path, target_directory):
         logger.debug('- Unpacking %s' % (path,))
@@ -399,15 +426,15 @@ class PackageManager(BasePackageManager):
                 return False
             return True
 
-    def read_package_requires_file(self, package_dir):
+    def read_package_requires_file(self, dist_dir):
         """Returns a list of dependencies for an unpacked package dir."""
-        name = os.listdir(package_dir)[0]
-        dist_dir = os.path.join(package_dir, name)
-        name, version = package_to_requirement(name).split('==')
         if not self.has_egg_info(dist_dir):
             return []
 
-        egg_info_dir = '{0}.egg-info'.format(name.replace('-', '_'))
+        files = os.listdir(dist_dir)
+        files = [f for f in files if f.endswith('egg-info')]
+        egg_info_dir = files[0]
+
         for dirpath, dirnames, _ in os.walk(dist_dir):
             if egg_info_dir in dirnames:
                 requires = os.path.join(dirpath, egg_info_dir,
@@ -432,13 +459,21 @@ class PackageManager(BasePackageManager):
         """
         logger.debug('- Extracting dependencies for %s' % (path,))
         with logger.indent():
-            build_dir = tempfile.mkdtemp()
-            unpack_dir = os.path.join(build_dir, 'build')
-            try:
-                self.unpack_archive(path, unpack_dir)
-                deps = self.read_package_requires_file(unpack_dir)
-            finally:
-                shutil.rmtree(build_dir)
+            if os.path.isdir(path):
+                # this is a directory in case if path is pointing to
+                # VCS checkout
+                deps = self.read_package_requires_file(path)
+            else:
+                build_dir = tempfile.mkdtemp()
+                unpack_dir = os.path.join(build_dir, 'build')
+                try:
+                    self.unpack_archive(path, unpack_dir)
+                    name = os.listdir(unpack_dir)[0]
+                    dist_dir = os.path.join(unpack_dir, name)
+
+                    deps = self.read_package_requires_file(dist_dir)
+                finally:
+                    shutil.rmtree(build_dir)
         logger.debug('Found: %s' % (deps,))
         return deps
